Fix broken Sloth tests
